polygons/requirements.txt
Django==3.2.9
django-rest-swagger==2.2.0
djangorestframework==3.12.4
psycopg2-binary==2.9.3
psycopg2==2.9.3


